% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ratio_funs.R
\name{add_ratios_to_df}
\alias{add_ratios_to_df}
\title{df data.frame with columns corresponding to metabolites (or similar).
names [vector] of metabolite names.
ratioop one of c("subtract", "divide") -> use subtract to make ratios for logarithmic data.}
\usage{
add_ratios_to_df(df, names, ratioop = c("subtract", "divide"))
}
\description{
df data.frame with columns corresponding to metabolites (or similar).
names [vector] of metabolite names.
ratioop one of c("subtract", "divide") -> use subtract to make ratios for logarithmic data.
}

