% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ratio_funs.R
\name{compute_pgain}
\alias{compute_pgain}
\title{add p-gain column to output data.frame
R is data.frame with results of the tests.
Returns data.frame with additional column.}
\usage{
compute_pgain(R)
}
\description{
add p-gain column to output data.frame
R is data.frame with results of the tests.
Returns data.frame with additional column.
}

