% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipeline_overall.R
\name{pipeline_overall}
\alias{pipeline_overall}
\title{Group comparisons of metabolomics data.}
\usage{
pipeline_overall(data, groups, refcat, compname, vargroups = data.frame(),
  test = c("linreg", "fisher"), covars = data.frame(),
  ratio_args = set_ratio_args(), repetitions = 0,
  boxplot_args = set_boxplot_args2())
}
\arguments{
\item{data}{[data.frame] with metabolites as columns.}

\item{groups}{[vector] grouping variable with length matching the number of rows of \code{data}}

\item{compname}{[character] prefix for name of output files (excel file, boxplots, volcano plots).}

\item{vargroups}{[data.frame] with columns being grouping variables of the metabolites in \code{data}.
Used for computing aggregated z-scores for the groups to be compared instead of comparing the single metabolites.}

\item{test}{[character] what test is to be performed? Linear regression F-test (if no covariates provided, equivalent to Anova) or
fishers exact test for comparing the number of NA (below threshold concentration) vs. non-NA values per metabolite.}

\item{repetitions}{[numeric] Number of repetitions for computing resampling-corrected p-values. Usually, at least 200 are recommended,
though higher accuracy can only obtain with greater numbers of repetitions.}

\item{ratios}{[logical] Should ratios of the metabolites be computed an tested? Will quickly lead to long runtimes.}

\item{ratioop}{[character] Should ratios be computed by "subtract"ing the metabolite values (for logarithmic data) or by "divide"ing?}
}
\value{
Returns NULL, all output (excel file, pdf with boxplots and volcano plots) is saved into the working directory.
}
\description{
Computes Overall F-tests (equivalent) or fisher-tests, for all groups in all metabolites given in "data".
}

