% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipeline_overall_old.R
\name{pipeline_overall2}
\alias{pipeline_overall2}
\title{Pipeline for overall F-tests of multiple groups.}
\usage{
pipeline_overall2(data, groups, refcat, compname, vargroups = data.frame(),
  test = "linreg", covars = data.frame(), ratios = FALSE,
  ratioop = "subtract", ratio_cutoff = 0.1, repetitions = 0,
  boxplot_args = set_boxplot_args2())
}
\description{
Pipeline for overall F-tests of multiple groups.
}

