% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pipeline_pairwise.R
\name{pipeline_pairwise}
\alias{pipeline_pairwise}
\title{Pairwise group comparisons of metabolomics data.}
\usage{
pipeline_pairwise(data, groups, compname, vargroups = data.frame(),
  covars = data.frame(), test = c("linreg", "fisher"),
  ratio_args = set_ratio_args(), repetitions = 0,
  boxplot_args = set_boxplot_args(), volc_args = set_volc_args(),
  no_plots = FALSE, verbose = TRUE, ncores = 1)
}
\arguments{
\item{data}{[data.frame] with metabolites as columns.}

\item{groups}{[vector] grouping variable with length matching the number of rows of \code{data}}

\item{compname}{[character] prefix for name of output files (excel file, boxplots, volcano plots).}

\item{vargroups}{[data.frame] with columns being grouping variables of the metabolites in \code{data}.
Used for computing aggregated z-scores for the groups to be compared instead of comparing the single metabolites.}

\item{covars}{[data.frame] with variables to be controlled for.}

\item{test}{[character] what test is to be performed? Linear regression (if no covariates provided, equivalent to t-test) or
fishers exact test for comparing the number of NA (below threshold concentration) vs. non-NA values per metabolite.}

\item{repetitions}{[numeric] Number of repetitions for computing resampling-corrected p-values. Usually, at least 200 are recommended,
though higher accuracy can only obtain with greater numbers of repetitions.}

\item{boxplot_args}{see \link{\code{set_boxplot_args}}}

\item{volc_args}{see \link{\code{set_volc_args}}}

\item{no_plots}{[logical] Should all plots be omitted? This will speed up the function considerably.}

\item{verbose}{[logical] Should progress be printed?}

\item{ncores}{[numeric] number of CPUs to be used. If >1, computations will be parallelized.}

\item{ratios}{[logical] Should ratios of the metabolites be computed an tested? Will quickly lead to long runtimes.}

\item{ratioop}{[character] Should ratios be computed by "subtract"ing the metabolite values (for logarithmic data) or by "divid"ing?}
}
\value{
Returns NULL, all output (excel file, pdf with boxplots and volcano plots) is saved into the working directory.
}
\description{
Computes pairwise t-tests (or linear regressions when covariates present) or fisher-tests, of the difference of all groups in all metabolites given in "data".
}

