% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{print_boxplot}
\alias{print_boxplot}
\title{Plotting function for pipeline_pairwise}
\usage{
print_boxplot(X, pvalue, name, boxplot_args)
}
\description{
takes data.frame with results (from compare_grps function)
takes data.frame without NAs
}

