% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{print_boxplot2}
\alias{print_boxplot2}
\title{Plotting function for pipeline_overall}
\usage{
print_boxplot2(X, name, p_overall, boxplot_args)
}
\description{
Plotting function for pipeline_overall
}

