% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{print_volc}
\alias{print_volc}
\title{Takes data.frame with results from compare_grps function
and volc_args: list with parameters that is set by function set_volc_args()
prints and returns the plot}
\usage{
print_volc(datframe, volc_args)
}
\description{
Takes data.frame with results from compare_grps function
and volc_args: list with parameters that is set by function set_volc_args()
prints and returns the plot
}

