% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_args_funs.R
\name{set_boxplot_args}
\alias{set_boxplot_args}
\title{Set arguments for boxplots in \link{\code{pipeline_pairwise}}}
\usage{
set_boxplot_args(sort_boxplots = FALSE, ylim = NULL, p_cutoff = Inf,
  ggplot = TRUE, colors = c("coral2", "turquoise3"), fill = "white",
  size = 1, shape = 21, alpha = 1)
}
\arguments{
\item{sort_boxplots}{should boxplots be sorted after p-values?}

\item{ylim}{[numeric vector(length = 2)] set ylim for all boxplots.}

\item{p_cutoff}{[numeric] Only metabolites with below cut-off p-value will be shown in boxplots. Default is no cut-off.}

\item{ggplot}{should ggplot2 package be used for plotting?}

\item{colors}{colors}

\item{fill}{fillcolor}

\item{size}{size of the points}

\item{shape}{shape}

\item{alpha}{transparency}
}
\description{
Set arguments for boxplots in \link{\code{pipeline_pairwise}}
}

