% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_args_funs.R
\name{set_boxplot_args2}
\alias{set_boxplot_args2}
\title{Set arguments for boxplots in \link{\code{pipeline_overall}}}
\usage{
set_boxplot_args2(sort_boxplots = FALSE, no_plots = FALSE, ylim = NULL,
  p_cutoff = Inf, colors = c("#E69F00", "#56B4E9", "#009E73"),
  ggplot = TRUE, fill = "white", size = 1, shape = 21, alpha = 1)
}
\arguments{
\item{sort_boxplots}{should boxplots be sorted after p-values?}

\item{no_plots}{[logical] Should all plots be omitted? This will speed up the function considerably.}

\item{ylim}{[numeric vector(length = 2)] set ylim for all boxplots.}

\item{p_cutoff}{[numeric] Only metabolites with below cut-off p-value will be shown in boxplots. Default is no cut-off.}

\item{colors}{colors}

\item{ggplot}{should ggplot2 package be used for plotting?}

\item{fill}{fillcolor}

\item{size}{size of the points}

\item{shape}{shape}

\item{alpha}{transparency}
}
\description{
Set arguments for boxplots in \link{\code{pipeline_overall}}
}

