% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_args_funs.R
\name{set_volc_args}
\alias{set_volc_args}
\title{Set arguments for volcano plots \link{\code{pipeline_pairwise}} and \link{\code{pipeline_continuous}}.}
\usage{
set_volc_args(cutoff = c(beta = 0.5, p = 0.05), cutoff_fdr = FALSE,
  colors = c("coral2", "turquoise3"), size = 1.75, shape = 21,
  fill = "white", alpha = 1)
}
\arguments{
\item{cutoff}{[numeric vector] Set cut-off values for the coefficient and p-value. Metabolites with values beyond the cut-off will be colored and labeled.}

\item{cutoff_fdr}{[logical] Should the FDR corrected p-values (instead of the naive, uncorrected) be used as cutoff?}

\item{colors}{colors}

\item{size}{size of the points}

\item{shape}{shape}

\item{fill}{fillcolor}

\item{alpha}{transparency}
}
\description{
Set arguments for volcano plots \link{\code{pipeline_pairwise}} and \link{\code{pipeline_continuous}}.
}

